refactor: polish match() PR — constant, docs, test consolidation
- Extract GqlMatchPlan.ANON_VAR_PREFIX constant to replace the raw
  "$anon" string literal scattered across DefaultGqlPlanner and
  GqlMatchStep
- Add upgrade guide NOTE that match(Traversal...) is deprecated as of
  4.0.0 in favour of match(String)
- Clarify mid-traversal match() semantics in reference docs: the full
  graph is scanned independently per incoming traverser, not joined on
  the traverser's current element
- Consolidate the path-label guard tests into the existing
  GqlMatchStepTest in ...step.map, removing the duplicate top-level
  class; guard tests now use a try-with-resources modern graph
